+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Support $refs to load markdown files into the spec and render them appropriately.
 * Correctly show deprecated arrays and hide deprecated elements from examples.
 * Add aria labels and roles to `<select>` and `<table>` elements.
+* [Bug] - don't force adding the spec url to the list of servers if the server list is already populated
 
 ## 0.7 ##
 * Introduce `setAuthenticationConfiguration` to set any all configuration related to token management, including clientIds and explicit tokens.